Add NewRequest component for submitting claims with form validation and popups
